# openspec/openapi-v1.yaml
openapi: 3.1.0
info:
  title: AstrBot OpenAPI v1
  version: 0.1.0
  description: >
    Target REST contract for migrating AstrBot HTTP APIs to /api/v1.
    Dynamic AstrBot configuration payloads are intentionally modeled as open
    JSON objects because their schemas are provided at runtime by template
    endpoints.
servers:
  - url: http://localhost:6185/api/v1
    description: Local AstrBot server
security:
  - ApiKeyAuth: []
...
paths:
...
  /commands:
    get:
      tags: [Commands]
      summary: List plugin commands
      operationId: listCommands
      x-astrbot-scope: plugin
      parameters:
        - name: config_id
          in: query
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /commands/{command_id}:
    patch:
      tags: [Commands]
      summary: Update command enabled state, alias, or permission group
      operationId: updateCommand
      x-astrbot-scope: plugin
      parameters:
        - $ref: "#/components/parameters/CommandId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommandPatchRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
  /tools:
    get:
      tags: [Tools]
      summary: List LLM tools
      operationId: listTools
      x-astrbot-scope: tool
      parameters:
        - name: origin
          in: query
          schema:
            type: string
            enum: [builtin, plugin, mcp]
        - name: enabled
          in: query
          schema:
            type: boolean
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /tools/{tool_id}/enabled:
    patch:
      tags: [Tools]
      summary: Enable or disable an LLM tool
      operationId: setToolEnabled
      x-astrbot-scope: tool
      parameters:
        - $ref: "#/components/parameters/ToolId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnabledPatch"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /mcp/servers:
    get:
      tags: [MCP]
      summary: List MCP servers
      operationId: listMcpServers
      x-astrbot-scope: tool
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    post:
      tags: [MCP]
      summary: Add an MCP server
      operationId: createMcpServer
      x-astrbot-scope: tool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/McpServerConfig"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /mcp/servers/by-name:
    put:
      tags: [MCP]
      summary: Update an MCP server by body name
      operationId: updateMcpServerByName
      x-astrbot-scope: tool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [server_name]
              properties:
                server_name:
                  type: string
                config:
                  $ref: "#/components/schemas/DynamicConfig"
                enabled:
                  type: boolean
              additionalProperties: true
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [MCP]
      summary: Delete an MCP server by query name
      operationId: deleteMcpServerByName
      x-astrbot-scope: tool
      parameters:
        - name: server_name
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /mcp/servers/enabled:
    patch:
      tags: [MCP]
      summary: Enable or disable an MCP server by body name
      operationId: setMcpServerEnabledByName
      x-astrbot-scope: tool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [server_name, enabled]
              properties:
                server_name:
                  type: string
                enabled:
                  type: boolean
              additionalProperties: false
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /mcp/servers/test:
    post:
      tags: [MCP]
      summary: Test an MCP server connection by body name
      operationId: testMcpServerByName
      x-astrbot-scope: tool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [server_name]
              properties:
                server_name:
                  type: string
                mcp_server_config:
                  $ref: "#/components/schemas/DynamicConfig"
                config:
                  $ref: "#/components/schemas/DynamicConfig"
              additionalProperties: true
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /mcp/servers/{server_name}:
    put:
      tags: [MCP]
      summary: Update an MCP server
      operationId: updateMcpServer
      x-astrbot-scope: tool
      parameters:
        - $ref: "#/components/parameters/ServerName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/McpServerConfig"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [MCP]
      summary: Delete an MCP server
      operationId: deleteMcpServer
      x-astrbot-scope: tool
      parameters:
        - $ref: "#/components/parameters/ServerName"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /mcp/servers/{server_name}/enabled:
    patch:
      tags: [MCP]
      summary: Enable or disable an MCP server
      operationId: setMcpServerEnabled
      x-astrbot-scope: tool
      parameters:
        - $ref: "#/components/parameters/ServerName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnabledPatch"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /mcp/servers/{server_name}/test:
    post:
      tags: [MCP]
      summary: Test an MCP server connection
      operationId: testMcpServer
      x-astrbot-scope: tool
      parameters:
        - $ref: "#/components/parameters/ServerName"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                mcp_server_config:
                  type: object
                  additionalProperties: true
              additionalProperties: true
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /mcp/providers/modelscope/sync:
    post:
      tags: [MCP]
      summary: Sync MCP servers from ModelScope
      operationId: syncModelScopeMcpServers
      x-astrbot-scope: tool
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelScopeSyncRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /skills:
    get:
      tags: [Skills]
      summary: List skills
      operationId: listSkills
      x-astrbot-scope: skill
      parameters:
        - name: enabled
          in: query
          schema:
            type: boolean
        - name: source
          in: query
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    post:
      tags: [Skills]
      summary: Upload or import a skill archive
      operationId: uploadSkill
      x-astrbot-scope: skill
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/SkillUploadRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /skills/batch:
    post:
      tags: [Skills]
      summary: Upload multiple skill archives
      operationId: uploadSkillsBatch
      x-astrbot-scope: skill
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /skills/by-name:
    patch:
      tags: [Skills]
      summary: Update skill metadata or enabled state by body name
      operationId: updateSkillByName
      x-astrbot-scope: skill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [skill_name]
              properties:
                skill_name:
                  type: string
                enabled:
                  type: boolean
                active:
                  type: boolean
              additionalProperties: true
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Skills]
      summary: Delete a skill by query name
      operationId: deleteSkillByName
      x-astrbot-scope: skill
      parameters:
        - name: skill_name
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /skills/archive:
    get:
      tags: [Skills]
      summary: Download a skill archive by query name
      operationId: downloadSkillByName
      x-astrbot-scope: skill
      parameters:
        - name: skill_name
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Skill archive
          content:
            application/zip:
              schema:
                type: string
                format: binary
  /skills/files:
    get:
      tags: [Skills]
      summary: List files in a skill by query name
      operationId: listSkillFilesByName
      x-astrbot-scope: skill
      parameters:
        - name: skill_name
          in: query
          required: true
          schema:
            type: string
        - name: path
          in: query
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /skills/file:
    get:
      tags: [Skills]
      summary: Get skill file content by query name and path
      operationId: getSkillFileByName
      x-astrbot-scope: skill
      parameters:
        - name: skill_name
          in: query
          required: true
          schema:
            type: string
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Text"
    put:
      tags: [Skills]
      summary: Update skill file content by body name and path
      operationId: updateSkillFileByName
      x-astrbot-scope: skill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [skill_name, path, content]
              properties:
                skill_name:
                  type: string
                path:
                  type: string
                content:
                  type: string
              additionalProperties: false
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /skills/{skill_name}:
    patch:
      tags: [Skills]
      summary: Update skill metadata or enabled state
      operationId: updateSkill
      x-astrbot-scope: skill
      parameters:
        - $ref: "#/components/parameters/SkillName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillPatchRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Skills]
      summary: Delete a skill
      operationId: deleteSkill
      x-astrbot-scope: skill
      parameters:
        - $ref: "#/components/parameters/SkillName"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /skills/{skill_name}/archive:
    get:
      tags: [Skills]
      summary: Download a skill archive
      operationId: downloadSkill
      x-astrbot-scope: skill
      parameters:
        - $ref: "#/components/parameters/SkillName"
      responses:
        "200":
          description: Skill archive
          content:
            application/zip:
              schema:
                type: string
                format: binary
  /skills/{skill_name}/files:
    get:
      tags: [Skills]
      summary: List files in a skill
      operationId: listSkillFiles
      x-astrbot-scope: skill
      parameters:
        - $ref: "#/components/parameters/SkillName"
        - name: path
          in: query
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /skills/{skill_name}/files/{file_path}:
    get:
      tags: [Skills]
      summary: Get skill file content
      operationId: getSkillFile
      x-astrbot-scope: skill
      parameters:
        - $ref: "#/components/parameters/SkillName"
        - $ref: "#/components/parameters/FilePath"
      responses:
        "200":
          $ref: "#/components/responses/Text"
    put:
      tags: [Skills]
      summary: Update skill file content
      operationId: updateSkillFile
      x-astrbot-scope: skill
      parameters:
        - $ref: "#/components/parameters/SkillName"
        - $ref: "#/components/parameters/FilePath"
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /skills/neo/candidates:
    get:
      tags: [Skills]
      summary: List Shipyard Neo skill candidates
      operationId: listNeoSkillCandidates
      x-astrbot-scope: skill
      parameters:
        - name: skill_key
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Ok:
      description: Standard AstrBot success response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessEnvelope"
...
  parameters:
...
    CommandId:
      name: command_id
      in: path
      required: true
      description: URL-encoded command handler full name.
      schema:
        type: string
...
    ToolId:
      name: tool_id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    DynamicConfig:
      type: object
      additionalProperties: true
...
    EnabledPatch:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
      additionalProperties: false
...
    CommandPatchRequest:
      type: object
      properties:
        enabled:
          type: boolean
        alias:
          type: string
        aliases:
          type: array
          items:
            type: string
        permission_group:
          type: string
      additionalProperties: false
...